Add rough merge interface
