+++ camp_fin/settings.py
@@ -60,6 +60,7 @@ INSTALLED_APPS = [
     "camp_fin.apps.CampFinConfig",
     "pages.apps.PagesConfig",
     "ckeditor",
+    "django_select2",
 ]
 
 MIDDLEWARE = [
@@ -190,6 +191,7 @@ DATABASES["default"] = dj_database_url.parse(
 # Caching
 
 cache_backend = "dummy.DummyCache" if DEBUG is True else "db.DatabaseCache"
+cache_backend = "db.DatabaseCache"
 CACHES = {
     "default": {
         "BACKEND": f"django.core.cache.backends.{cache_backend}",
